Refact: repostiory 위치 수정 및 행정구역 repository 분할
기존에는 한 repostiory에서 원본 행정구역 테이블 생성과 분할 행정구역 테이블 생성을 같이 했기 때문에 이 둘을 분할
